requirements.txt
groq>=1.1.1
python-dotenv>=1.0.1
rich>=13.7.0
pydantic>=2.9.0
asyncio
